requirements.txt
beautifulsoup4==4.12.3
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
tenacity==9.0.0
pandas==2.2.3
urllib3<2
